requirements.txt
# Core testing
pytest==8.3.2
pytest-xdist==3.6.1
pytest-rerunfailures==14.0

# Selenium + drivers
selenium==4.24.0
webdriver-manager==4.0.2

# Reporting
allure-pytest==2.13.5

# Utilities
PyYAML==6.0.2
colorlog==6.8.2
psutil==6.0.0

# Screen recording helper (FFmpeg required on system PATH)
mss==9.0.2
Pillow==10.4.0
